For #18202 - Refactor BiometricPromptFeature out of logins into settings
