+++ app/src/test/java/org/mozilla/fenix/settings/biometric/ext/BiometricManagerKtTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.settings.logins.biometric.ext
+package org.mozilla.fenix.settings.biometric.ext
 
 import androidx.biometric.BiometricManager
 import androidx.biometric.BiometricManager.BIOMETRIC_ERROR_HW_UNAVAILABLE
